Teach scss-lint about nested property syntax
Prior to this commit, scss-lint was giving false positives for nested
properties:

  http://sass-lang.com/docs/yardoc/file.SASS_REFERENCE.html#nested_properties

We have some of these in our code base, so this commit teaches scss-lint
to grok them.

As the regex was starting to get gnarly, I broke it into multi-line
syntax and added comments. I also replaced the negative look-behind with
a simpler positive construction.

Related Story:
  http://www.pivotaltracker.com/story/show/41761797 (delivered)
  ("Teach scss-lint about nested property syntax")

Change-Id: I27dde8025fb43b383aadd4c625429c9e3f4d5d7c
Reviewed-on: https://gerrit.causes.com/16962
Tested-by: Greg Hurrell <greg@causes.com>
Reviewed-by: Lann Martin <lann@causes.com>