Bump version to v0.5.2
The previous commit which deals with not commit untracked files still
picks up changes to tracked files. Unfortunately, I had a modification
to a file that was _already_ tracked, so it was accidentally included in
the gem that was build.

Upgrading the version allows us to push a new version to fix this
problem. I've learned a few painful lessons today about building gems.

Change-Id: I327925360007288d58f77f8f011a0e005039c80e
Reviewed-on: https://gerrit.causes.com/17803
Reviewed-by: Tony Wooster <tony.wooster@causes.com>
Tested-by: Shane da Silva <shane@causes.com>