Fix LinterRegistry.extract_linters_from
This wasn't working because we weren't properly scoping the constant
lookup to the `Linter` namespace. We didn't catch this in the specs
because the specs were declaring their one-off classes without any
namespaces, so they _appeared_ to be correct.

The solution was to scope the `const_get` lookup to the `Linter`
namespace, and update the specs to actually create linters in the
`SCSSLint::Linter` namespace.

Fixes #13

Change-Id: I07719f4b457b1d8886784fcefb93ae1660ad82fc
Reviewed-on: https://gerrit.causes.com/22079
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Kim Douglas Clarence Mason <kim.mason@causes.com>
Reviewed-by: Aiden Scandella <aiden@causes.com>