Flesh out the readme
I just linked to scss-lint on Hacker News:

  https://news.ycombinator.com/item?id=5785885

and realized that the README doesn't really do a good job of telling
people the details of what it does (short of reading the source code or
trying it out).

This commit adds some text describing the rationale of the tool (ie.
that it is opinionated, and that consistency is more important than the
superficial details) and a list of some of the lints that are included.

As a bonus, includes a link to overcommit, which is related.

Change-Id: I5ac859b01bc9a495420bb92f160819553211b483
Reviewed-on: https://gerrit.causes.com/23023
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Aiden Scandella <aiden@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>