Use Engine filename in add_lint
In preparation for a commit that adds support for traversing
`Sass::Script::Node`s on top of `Sass::Tree::Node`s, the dependency on
fetching the filename corresponding to the lint from the node must be
removed, since `Sass::Script::Node` instances do not have a `filename`
method.

Since the engine will have the filename passed in, use that instead.

Change-Id: I6040f69e997d59755b608cc6e3d22699f0014f80
Reviewed-on: https://gerrit.causes.com/23137
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>