Consolidate linter test setup code
Each of the linter tests did the exact same thing, namely:

  * Initialize a parse tree with the CSS for the given example
  * Run the linter against that parse tree

Since linter tests will always have this format, we can DRY up this
setup code by moving it into the spec_helper file as a global `before`
hook.

By checking within the hook whether the current `described_class` is a
subclass of linter, we can dynamically choose whether to perform the
linter setup or not.

Change-Id: I7a8af2c4be6fb4abd02cd0518b2768853e18c7c8
Reviewed-on: https://gerrit.causes.com/23173
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>