Upgrade Sass gem dependency 3.2.8 -> 3.2.9
A bug reported in #19 had a stack trace referencing version 3.2.9 of the
`sass` gem. This was somewhat surprising since this project's
Gemfile.lock file was coded to 3.2.8.

However, it is ultimately the gemspec file that must encode the version
dependencies for a gem. For most gems, this wouldn't be set to an exact
version number, and would instead encode the pessimistic version
constraint (i.e. '~> 3.2').

Unfortunately, since `scss-lint` is tightly coupled with the
implementation of the underlying `sass` gem due to how it monkey-patches
the parse tree, we need to enforce an exact version here so that we
don't get bug reports from users using different versions of
`sass` which may have modified the implementation of the parse tree.

This hard-codes the dependency to version 3.2.9 of the sass gem.

Change-Id: I04f62a385fb65ddf2b0a53ddb1af71b0ec8212d5
Reviewed-on: https://gerrit.causes.com/23259
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>