Handle sorting of vendor-prefixed properties
Previously, the `SortedProperties` linter would not handle
vendor-prefixed properties in any special way, resulting in them always
being required to be at the beginning of the list of properties (since
they all start with a hyphen).

A more appropriate solution would be to enforce an ordering where
vendor-prefixed properties are sorted by vendor, and placed before the
associated non-prefixed property. For example:

    border: 0;
    -moz-border-radius: 3px;
    -o-border-radius: 3px;
    -webkit-border-radius: 3px;
    border-radius: 3px;
    color: #ccc;

This commit adds support for this kind of ordering.

In the process, the line being reported is now the line of the first
property that is out of order (rather than just the first property in
the list of properties, as was previously the case).

Note that this implementation still ignores interpolated variables.

Closes #22

Change-Id: I1d3d38d11eaf8a13c79b7cfacb0130f5c6e3873a
Reviewed-on: https://gerrit.causes.com/27824
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Lann Martin <lann@causes.com>