Update Gemfile.lock to fix TravisCI build
Normally we haven't had to update the `Gemfile.lock` file since when we
run tests locally they work as expected.

However, going forward we need to ensure we update this file by running
`bundle install` locally when we bump the gem version, otherwise the
TravisCI build will fail trying to find to find the outdated version of
`scss-lint`.

Change-Id: I8cacfd7f8b98558a1afe22d39a4878c58fdaf03f
Reviewed-on: https://gerrit.causes.com/27862
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>