Change CapitalizationInSelector to use SelectorVisitor
The previous implementation of `CapitalizationInSelector` was flawed in
that it would incorrectly report a lint for a selector like:

    [data-text=someText] { ... }

This was due to the fact that we were not differentiating between the
attribute name and its value in the selector.

Changing the linter to use the new selector-based `visit_*` methods
allows us to get around this problem.

Fixes #40

Change-Id: Id5ec24121b44d98376f3e231caf5c4b95356e337
Reviewed-on: https://gerrit.causes.com/28830
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>