Make UsageName lints more specific
Change the descriptions of lints reported by the `UsageName` linter to
be more-specific to the context of the problem (e.g. specify whether the
problematic name is for a function, variable, mixin, etc.).

In the process I was able to remove a method in the `Utils` module into
the only location it was used (the `UsageName` linter).

Change-Id: Id1fff3a523cc252c78023fb9b2391d55f490a836
Reviewed-on: https://gerrit.causes.com/28833
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>