Ignore color keywords in HexFormat linter
The `HexFormat` linter would fail on CSS color keywords, e.g. `gold` or
`orange`. This was due to the fact that the `original` value of the
`Color` object would not be set for colors specified via keyword. This
would only happen for color keywords that appear in Sass script however
(not color keywords that appear as normal property values, as those
don't get created as `Color` objects by the parser).

The solution was to check that `original` was defined before attempting
to run the lint.

Change-Id: I10d87e497d60eee642c4ed9cd6347a3453138147
Reviewed-on: https://gerrit.causes.com/29013
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>