Fix LeadingZero linter to work with Sass script
The initial implementation of the `LeadingZero` linter did not include
support for finding leading zeros in numeric Sass script values--it only
had support for detecting leading zeros in property values.

In order to add support for this, we (yet again) have to monkey-patch
the Sass lexer in order to store the original string for the numeric
value. We will hopefully be able to merge support for this upstream, but
for now this allows us to fix the bug.

In the process of adding this new support, it was noticed that the code
for the `remove_quoted_strings` method of the `ColorKeyword` linter was
repeated in the `LeadingZero` linter. Thus the `remove_quoted_strings`
method was extracted into the `Utils` module so it could be used by both
linters.

Change-Id: Id1d057ea8e97bab7aa951b34d40bc5cc47764c36
Reviewed-on: https://gerrit.causes.com/29752
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Tom Dooner <tom.dooner@causes.com>