Add linter documentation to README
Long overdue is an overhaul to the README to actually explain what is
getting linted and why. This is more useful to developers as they tend
to quickly scan the README before deciding whether to use a gem.

Going forward it should be a lot easier to keep this up-to-date.

Change-Id: I4383b32a3247f993bd709ca940c1c01ac5d89f02
Reviewed-on: https://gerrit.causes.com/30683
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>