Load config file based on file being linted
Add support for automatically loading a configuration file based on the
path of the file being linted. The semantics here are similar to Rubocop
in that for a file `/path/to/file.scss`, we will check in order:

  1. `/path/to/.scss-lint.yml`
  2. `/path/.scss-lint.yml`
  3. `/.scss-lint.yml`

...and stop on the first file found.

This gives the developer a lot of power in configuring `scss-lint` on a
per-directory basis.

Change-Id: I0eb5c11891cc4709b4e7b9f8c4ff07e742b8893a
Reviewed-on: https://gerrit.causes.com/30922
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>