Extract option parser initialization
The `parse_arguments` method of the `CLI` class was rather long due to
the inline initialization of the `OptionParser`.

Extracting this initialization into a separate method makes it clearer
that the purpose of the `parse_arguments` method is to simply parse
arguments and handle invalid arguments if they are provided.

Change-Id: Icf095dc76226987c01fc7011e99db88c723eda12
Reviewed-on: https://gerrit.causes.com/31534
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>