Ignore interpolation in SingleLinePerSelector
The `SingleLinePerSelector` linter would incorrectly report lints when
selector contained interpolation with a comma in it, e.g.

    .my-#{function(1, 2)}-selector .item {
    }

The solution is to ignore the content of any interpolation, as it has no
bearing on what the linter is trying to check.

In the process I also updated the linter to not define a `description`
method, as this style of linter declaration should go away in favour of
explicitly passing messages to `add_lint`.

Fixes #67

Change-Id: Ice0191e0133bbe2490882c1dd4237fce12b41282
Reviewed-on: https://gerrit.causes.com/32594
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>