Add YARD to CLI, Lint, Linter, and Runner
I find code much more easy to work with when the argument datatypes and
return value datatypes for each method are clearly specified. This
commit attempts to add that documentation via YARD tags to a few of the
classes. I'm not entirely sure I got everything correct, but it seems
like a pretty good start.

While I was at it, I aligned some code to improve readability.

Change-Id: I8ca68c5e05ec53a07497a94c12eade15c251b97c
Reviewed-on: http://gerrit.causes.com/36105
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>