Fix Indentation linter with at-root inline selectors
Due to the way the Sass parser creates the parse tree for `@at-root`
directives, the following two pieces of code produce the same parse
tree:

    @at-root {
      .something {
        ...
      }
    }

    @at-root .something {
      ...
    }

This was resulting in erroneous lints being reported in the
`Indentation` linter, since it couldn't tell which case was which given
the parse tree.

We can use a nasty hack of checking if the end position of the
`@at-root` node matches the start position of its first child, in which
case we know the child is inline instead of being properly nested.

Change-Id: Iaba01f67c5c3f4ac65b95b48e1c950a662eb227c
Reviewed-on: http://gerrit.causes.com/38085
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>