Fix TrailingSemicolon linter for `@include`s with blocks
This fixes https://github.com/causes/scss-lint/issues/198

A recent commit [1] generalized the TrailingSemicolon linter to also
check `@extend` and `@include`. This commit did not take into account
`@include`s with a block. @Steffen185 provided a simple repro case:

  $ cat test.scss
  body {
    @include foo {
      background: #f00;
    }
  }
  $ scss-lint test.scss
  test.scss:2 [W] TrailingSemicolon: Declaration should not have a space
  before the terminating semicolon

I took this case and added it to the spec for the TrailingSemicolon
linter. The fix included checking the `@include` for children, and in
such case continue checking children. There was already code for this
scenario for props, so I just refactored a bit to make use of the same
solution.

[1]: https://github.com/causes/scss-lint/commit/58689b1d18c7449509168f

Change-Id: If8f178d7b36a048bb7c00e75b0738b9efe0530ed
Reviewed-on: http://gerrit.causes.com/40988
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>