Upgrade sass dependency to 3.4.x series
Upgrade the minimum required `sass` gem version to 3.4.x. This adds
support for the latest Sass version.

Most of the fixes are small adjustments to to tiny changes in the parse
tree representation.

One major change is in the `SingleLinePerSelector` linter. Due to an
upstream change in the Sass parser, it broke the parsing of trailing
comments in rule selectors. However, according to
https://github.com/sass/sass/issues/1416, they will not fix the bug, as
from their perspective Sass does not follow semantic versioning and
makes no guarantees about its internal parse tree representation.

Thus we changed `SingleLinePerSelector` to use a cleaner method of
checking for newlines by inspecting the `CommaSequence` directly (which
is what we should have done a while ago, so it's good that this prompted
us to do it).

Change-Id: Ib71e4449270beb1d55ee236615979f574dd8a593
Reviewed-on: http://gerrit.causes.com/42246
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>