Fix SpaceAfterPropertyName linter with interpolation
A strange bug with the Sass parser would result in incorrect lints being
reported for properties following selectors with interpolation on a
single line.

A dirty workaround is to simply find the first colon in the source range
and check from that point to see if there is a space.

Change-Id: I2006a252c660ed52284e0f710caed5322689db42
Reviewed-on: http://gerrit.causes.com/43582
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>