Convert ControlComments module into helper class
Using a module for this functionality is a bit of a smell as we're
simply pulling out the functionality into a separate file rather than
isolating concerns.

Changing this class to a helper class removes the dependency on
understanding the `Linter` class internals and instead only interacts
with the `Linter` class's public interface.

It also removes the need for multiple invocations of the
`initialize_vars` method, as we can now do that one time in the
constructor.

Change-Id: Ie6a2efb5921d6f13688e8b5198ebfe614584d7d0
Reviewed-on: http://gerrit.causes.com/44953
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>