Add .projections.json configuration file
Editor plugins, like @tpope's vim-projectionist, can make use of this
configuration file to enable easy jumping between "alternate" files,
among other features. For example, with this configuration and
vim-projectionist installed, you can now be editing a lint and execute
`:A` to jump to that lint's spec file. This will save some time when
working with this project.

https://github.com/tpope/vim-projectionist

Change-Id: Ic7f96d5eb4398c0ee011929bce640a195ee204d3
Reviewed-on: http://gerrit.causes.com/45430
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>