Extract file finding into separate helper class
The CLI class is complicated. Simplify it by extracting the file-finding
concern into a separate helper class.

As a result, we now support `scss-lint` accepting file paths that do not
end in `.scss` or `.css`. If you specify `scss-lint test.txt` for
example, it will attempt to lint the file (and return a syntax error if
it cannot be parsed).

Change-Id: I5a3ae40cc2c99e5c0e4a0e915e8a433e93b62244
Reviewed-on: http://gerrit.causes.com/46308
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>