Fix ColorVariable to not report "transparent"
The keyword `transparent` typically isn't used as a color but more as a
way to remove colors.

Normally it was already excluded thanks to the `color?` helper method,
but for cases where `transparent` appeared in a function call or list
literal it would still report the offence. Fix it by checking `color?`
first.

Change-Id: I201c1babb6e8ee54b64ab73321bfe1879956557e
Reviewed-on: http://gerrit.causes.com/47200
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>