Demote PropertyUnits instance variables to local variables
There was no need to make these instance variables.

While converting them to locals I removed an unnecessary helper that
converted hyphens to underscores for the property name in the
configuration. Since YAML keys can have hyphens, there was no need for
this, and it would probably result in more confusion since developers
would expect to be able to write `line-height`, not `line_height`.