Simplify PropertyUnits implementation
Reduce the number of moving parts in the implementation to use fewer
helper methods. This results in fewer lines of code with a simpler
conceptual approach.

In the process, added some documentation and improved the error message
to include the actual list of units that are allowed, so developers
don't need to consult their configuration.