Fix broken test on Ruby 2.2.0+
As of Ruby 2.2.0, the sort order does not seem to be the same. Given
this affects a single test and does not affect the behavior of the
linter in a negative way (it's still reporting an out-of-order error),
I'm just going to relax the test to not explicitly define the line
number reported.