Fix DeclarationOrder false positives
As of Ruby 2.2.0 the sorting algorithm appears to have changed, which
results in an unstable sort of elements. These causes a false positive
since two nodes of the same type could have their order shuffled.

Fix this by including the index of the node in the sort so we preserve
order explicitly.