Reset lint count on each Linter#run
Not resetting the lint count on each run resulted in strange behavior
where depending on the order files were scanned you could get lints
in other files being filtered out.

This was because we were extracting lints at the end. We should extract
them as we go, returning a separate list per invocation of `Linter#run`.