Ensure `@else` statements are assigned a node_parent
The `Indentation` linter was crashing linting `@else` statements when
the `allow_non_nested_indentation` option was set.

This was due to the node not having its `node_parent` assigned. Fix it
by adding the `else` property of the node to the list of children in the
monkey patching of `IfNode`.