Add placeholder, -moz, -ms, and -webkit to PseudoElement
I noticed that this linter didn't include these, which are currently
giving false positives on my codebase.

I found the list of -ms-* pseudo-elements at
https://msdn.microsoft.com/en-us/library/windows/apps/hh767361.aspx
and -moz-* pseudo-elements at
https://developer.mozilla.org/en-US/docs/Web/CSS/Mozilla_Extensions

Unfortunately, I was unable to find a comprehensive source of -webkit-*
pseudo-elements, so I went with the couple that have been reported so
far.

Fixes #639