Add --color and --no-color options
We want to allow people to force the use or non-use of color when
running scss-lint. This commit does this by extracting the logging into
a Logger class and adding `--color` and `--no-color` options.

For this code, I essentially brought over the same stuff that was done
in HamlLint, and split out coloring and logging into separate methods on
the Logger so that we could color strings without outputting them in our
reporters.

This removes the dependency on the Rainbow gem, which is nice. There was
a note in the readme about Windows compatibility that I'm not really
sure if it still applies or not, but since the wording and links to more
information was Rainbow gem specific I decided to just remove it--we can
add Windows information back in again if it comes up.

Addresses #364