Visit children in SingleLinePerProperty
With the following configuration:

```yaml
SingleLinePerProperty:
  enabled: true
  allow_single_line_rule_sets: false
```

And the following SCSS:

```scss
.test { padding: 0; margin: 0; }
```

The SingleLinePerProperty correctly reports a lint here. However, if
this is nested:

```scss
.testing {
  .test { padding: 0; margin: 0; }
}
```

it is quiet.

It seems that the Sass parser is not visiting the child rule here. I
don't know enough about how this is supposed to work to know if that is
expected or not, but it seems that I can make this work by adding a line
that will recurse the children.

Fixes #709