Relax rake dependency version to allow >= 0.9
I am working in a project that is running an old version of rake, and
for historical reasons, it is going to be very difficult to upgrade to
rake 10. I tried looking into why this was set at rake 10 and I didn't
really come up with much, so I am relaxing this version requirement a
little.

I tried this out locally and specs pass on 0.9.* but not on 0.8.*, so
that seems like a good place to draw the line.